+++ bindings/angular2/src/directives/ons-input.ts
@@ -16,8 +16,8 @@ import {
  * @directive OnsInput
  * @selector ons-input
  * @description
- *   [en]Angular 2 directive for `<ons-input>` component.[/en]
- *   [ja]`<ons-input>`要素のAngular 2ディレクティブです。[/ja]
+ *   [en]Angular directive for `<ons-input>` component.[/en]
+ *   [ja]`<ons-input>`要素のAngularディレクティブです。[/ja]
  * @example
  *   <ons-input [(value)]="value"></ons-input>
  */
@@ -50,7 +50,7 @@ export class OnsInput implements OnChanges, OnDestroy {
     this._boundOnChange = this._onChange.bind(this);
     this._element = _elementRef.nativeElement;
 
-    this._element.addEventListener('change', this._boundOnChange);
+    this._element.addEventListener('input', this._boundOnChange);
   }
 
   _onChange(event: any) {
@@ -73,7 +73,7 @@ export class OnsInput implements OnChanges, OnDestroy {
   }
 
   ngOnDestroy() {
-    this._element.removeEventListener('change', this._boundOnChange);
+    this._element.removeEventListener('input', this._boundOnChange);
 
     this._element = null;
   }